--- a/pretty/doc.go
+++ b/pretty/doc.go
@@ -20,6 +20,3 @@
 //
 // See the Reflect and Print examples for what the output looks like.
 package pretty
-
-// TODO:
-//   - Catch cycles