--- a/pretty/public.go
+++ b/pretty/public.go
@@ -25,6 +25,8 @@ import (
 	"github.com/kylelemons/godebug/diff"
 )
 
+var emptyset = map[uintptr]bool{}
+
 // A Config represents optional configuration parameters for formatting.
 //
 // Some options, notably ShortList, dramatically increase the overhead
@@ -95,7 +97,7 @@ func (cfg *Config) fprint(buf *bytes.Buffer, vals ...interface{}) {
 		if i > 0 {
 			buf.WriteByte('\n')
 		}
-		cfg.val2node(reflect.ValueOf(val)).WriteTo(buf, "", cfg)
+		cfg.val2node(reflect.ValueOf(val), emptyset).WriteTo(buf, "", cfg)
 	}
 }