--- a/pretty/reflect.go
+++ b/pretty/reflect.go
@@ -29,8 +29,7 @@ func isZeroVal(val reflect.Value) bool {
 	return reflect.DeepEqual(val.Interface(), z)
 }
 
-func (c *Config) val2node(val reflect.Value) node {
-	// TODO(kevlar): pointer tracking?
+func (c *Config) val2node(val reflect.Value, seen map[uintptr]bool) node {
 
 	if !val.IsValid() {
 		return rawVal("nil")
@@ -60,14 +59,21 @@ func (c *Config) val2node(val reflect.Value) node {
 		if val.IsNil() {
 			return rawVal("nil")
 		}
-		return c.val2node(val.Elem())
+		if val.Kind() == reflect.Ptr {
+			p := val.Pointer()
+			if present, _ := seen[p]; present {
+				return rawVal(fmt.Sprintf("[recursion on 0x%x]", p))
+			}
+			seen[p] = true
+		}
+		return c.val2node(val.Elem(), seen)
 	case reflect.String:
 		return stringVal(val.String())
 	case reflect.Slice, reflect.Array:
 		n := list{}
 		length := val.Len()
 		for i := 0; i < length; i++ {
-			n = append(n, c.val2node(val.Index(i)))
+			n = append(n, c.val2node(val.Index(i), seen))
 		}
 		return n
 	case reflect.Map:
@@ -75,7 +81,7 @@ func (c *Config) val2node(val reflect.Value) node {
 		keys := val.MapKeys()
 		for _, key := range keys {
 			// TODO(kevlar): Support arbitrary type keys?
-			n = append(n, keyval{compactString(c.val2node(key)), c.val2node(val.MapIndex(key))})
+			n = append(n, keyval{compactString(c.val2node(key, seen)), c.val2node(val.MapIndex(key), seen)})
 		}
 		sort.Sort(n)
 		return n
@@ -92,7 +98,7 @@ func (c *Config) val2node(val reflect.Value) node {
 			if c.SkipZeroFields && isZeroVal(field) {
 				continue
 			}
-			n = append(n, keyval{sf.Name, c.val2node(field)})
+			n = append(n, keyval{sf.Name, c.val2node(field, seen)})
 		}
 		return n
 	case reflect.Bool: