--- a/pretty/reflect_test.go
+++ b/pretty/reflect_test.go
@@ -18,6 +18,7 @@ import (
 	"fmt"
 	"net"
 	"reflect"
+	"strings"
 	"testing"
 	"time"
 )
@@ -203,4 +204,21 @@ func TestVal2node(t *testing.T) {
 			t.Errorf("%s: got %#v, want %#v", test.desc, got, want)
 		}
 	}
+}
+
+func TestRecursion(t *testing.T) {
+
+	type Recursive struct {
+		R *Recursive
+	}
+
+	r := &Recursive{}
+	r.R = r
+	x := Sprint(r)
+
+	if !strings.Contains(x, "recursion on") {
+		t.Errorf("Expected formatted string to contain 'recursion on', got: %q",
+			x)
+	}
+
 }