Hello up there. I was using godebug/pretty to pretty-print differences in tests and ran into infinite recursion in pretty.Compare() for datastructures with cycles. I looked around on github in godebug network and saw @pwaller already addressed this in 2013. So here are original patches by @pwaller which I rebased to recent master resolving minor conflicts along the way.

e.g. for the following example

```go
package main

import (
        "github.com/kylelemons/godebug/pretty"
)

type A struct {
        Ref *A
}

func main() {
        var a1, a2 A
        a1.Ref = &a2
        a2.Ref = &a1
        pretty.Print(a1)
}
```

previously the output was:

```
(neo) (g.env) kirr@deco:~/src/neo$ go run x.go
runtime: goroutine stack exceeds 1000000000-byte limit
fatal error: stack overflow

runtime stack:
runtime.throw(0x5094fc, 0xe)
        /home/kirr/src/tools/go/go/src/runtime/panic.go:605 +0x95
runtime.newstack(0x0)
        /home/kirr/src/tools/go/go/src/runtime/stack.go:1050 +0x6e1
runtime.morestack()
        /home/kirr/src/tools/go/go/src/runtime/asm_amd64.s:415 +0x86

goroutine 1 [running]:
reflect.packEface(0x4ed120, 0xc42000e038, 0x199, 0x0, 0x0)
        /home/kirr/src/tools/go/go/src/reflect/value.go:96 +0xfe fp=0xc440100348 sp=0xc440100340 pc=0x49ac4e
reflect.valueInterface(0x4ed120, 0xc42000e038, 0x199, 0x1, 0x0, 0x0)
        /home/kirr/src/tools/go/go/src/reflect/value.go:955 +0xe9 fp=0xc4401003a0 sp=0xc440100348 pc=0x49e479
reflect.Value.Interface(0x4ed120, 0xc42000e038, 0x199, 0x1, 0x0)
        /home/kirr/src/tools/go/go/src/reflect/value.go:925 +0x44 fp=0xc4401003e0 sp=0xc4401003a0 pc=0x49e364
github.com/kylelemons/godebug/pretty.(*Config).val2node(0x5a7dd0, 0x4ed120, 0xc42000e038, 0x199, 0xc42000e038, 0x199)
        /home/kirr/src/neo/src/github.com/kylelemons/godebug/pretty/reflect.go:40 +0x14bc fp=0xc4401007c0 sp=0xc4401003e0 pc=0x4c7c4c
github.com/kylelemons/godebug/pretty.(*Config).val2node(0x5a7dd0, 0x4d8280, 0xc42000e030, 0x196, 0x4d8280, 0xc42000e030)
        /home/kirr/src/neo/src/github.com/kylelemons/godebug/pretty/reflect.go:63 +0x94a fp=0xc440100ba0 sp=0xc4401007c0 pc=0x4c70da
github.com/kylelemons/godebug/pretty.(*Config).val2node(0x5a7dd0, 0x4ed120, 0xc42000e030, 0x199, 0xc42000e030, 0x199)
        /home/kirr/src/neo/src/github.com/kylelemons/godebug/pretty/reflect.go:95 +0xb62 fp=0xc440100f80 sp=0xc440100ba0 pc=0x4c72f2
github.com/kylelemons/godebug/pretty.(*Config).val2node(0x5a7dd0, 0x4d8280, 0xc42000e038, 0x196, 0x4d8280, 0xc42000e038)
        /home/kirr/src/neo/src/github.com/kylelemons/godebug/pretty/reflect.go:63 +0x94a fp=0xc440101360 sp=0xc440100f80 pc=0x4c70da
...
```

and with the patches it does not crash:

```
(neo) (g.env) kirr@deco:~/src/neo$ go run x.go
{Ref: {Ref: {Ref: [recursion on 0xc42000e038]}}}
```

So please apply.

Thanks beforehand,
Kirill