@pwaller 2017-08-16
> We need to confirm that they're okay with their commits being contributed to this project. Please have them confirm that here in the pull request.

I'm okay with it. 👍