@kylelemons 2017-08-19
The fundamental problem with the implementations that I've seen in the network are that they don't satisfy the property that pretty.Compare should return true if and only if the recursive structures are identically recursive.  For instance, if I have a circularly linked list with circumference 5 and one with a circumference 3, they should be different, and two different lists with circumference 5 (and identical values and cursor) should be identical.  They also often have the property that they also elide portions of a structure that appear twice, even if it does not actually involve recursion.  So, if you have a struct that simply points to the same value twice, one will be printed as a pointer and (incorrectly) marked as recursion.

No matter what representation you choose, you run into problems:

1. Print pointers:

       {Value: 1, Next: {Value: 2, Next: {Value: 3, Next: [recursion to 0xfaaa0400]}}}

    This is relatively ugly, though precise, but it means that you can't tell if that's `1 -> 2 -> 3 -> 1` or basically anything else. 

1. Don't print recursive values at all, use ... to indicate "this value has already been printed":

       {Value: 1, Next: {Value: 2, Next: {Value: 3, Next: ...}}}

    This is pretty, but it means that `1 -> 2 -> 3 -> 1` compares equally to `1 -> 2 -> 3 -> 2` which is not desirable.

1. Print "recursion IDs"

       <id=1> {Value: 1, Next: {Value: 2, Next: {Value: 3, Next: [recursion to <id=1>]}}}

    This is also relatively ugly, but is at least precise and, as long as you iterate over things like maps in a deterministic fashion, predictable.  The downside to this is that you have to not only detect cycles, but also mark the value to which you're referring when it first appears, which is a challenge to do if you actually want the behavior described above where you omit duplicate objects even if they aren't recursive. If you like this approach, I would encourage you to look at [go-spew](https://github.com/davecgh/go-spew) which, I believe, takes this approach.

 I specifically left recursive support out of the library so that you couldn't accidentally run afoul of these in your code, you simply couldn't use pretty.  I recognize that it's a flaw, and I would like to solve it, though I don't have a perfect solution quite yet.  I'm thinking that I might prefer (2) from above, but with something more like

       {Value: 1, Next: {Value: 2, Next: {Value: 3, Next: (recursive) {Value: 1, Next: {Value: 2, Next: {Value: 3, Next: ...}}}}}}}

That is, print each value at most twice, and then begin eliding the values with ... beyond that.  I think this allows you to see the actual structure, avoid infinite recursion, and still be able to distinguish between different constructions of recursive values that are not morphologically identical.