@kylelemons 2017-08-20
If you can, take a look at #18 and see if that works for you.

@pwaller you may be interested in this as well.