@navytux 2017-08-20
@kylelemons thanks for feedback. With your description I think I now better understand the requirement that Compare always works predictably with retaining its properties. Thanks also for you patches. 

After thinking a bit I became to conclusion that option #3 should be optimal from both:

- pretty output consumer point of view, and
- requirement of making Compare always work reliably  (as you do note so about RecursiveContext  giving no 100% unambigous results).

I also looked a bit at go-spew and I think by default I like pretty a bit more for simplicity maybe...

So on my side I've tried to come up with the patches for #3:

https://lab.nexedi.com/kirr/godebug/commit/6569ee76
https://lab.nexedi.com/kirr/godebug/commit/5658db6e

The code still distinuish cycles from plain cases where one object is simply referenced twice from different places and a test is there for it:

https://lab.nexedi.com/kirr/godebug/commit/6569ee76d6#ce9925585901aa878fe4013faa1d39216185a646_228_226

That means there are 2 possibilities:

- either sleepy me is still missing something, or
- we should be ok with the changes to go.

I've also prepared all the patches as a pull-request for convenience: #19.

Thanks beforehand,
Kirill