@pwaller 2017-08-21
@kylelemons Nice thoughtful analysis, thanks for sharing. Solution looks great.