@kylelemons 2017-08-22
@navytux Thanks for the pr; I decided to experiment with (3) as well, and I agree that it works out better, though I'm still toying with exactly what representation I want for the references.  My implementation is slightly different from yours: I decided to fix stable map ordering so that I could assign the IDs at recursion time instead of assigning them at output time.

Take a look at #18 again and let me know what you think.