@navytux 2017-08-22
@kylelemons thanks for feedback. It is good to try to allocate the ids at recursion time though it still allocates ids in the order they are referenced, not in the order targets should be defined. For example with

```go
type A struct {
        X int
        *B
}

type B struct {
        *B
        *A
}

        a := &A{X: 777}
        b := &B{}
        a.B = b
        b.B = b
        b.A = a
```

my implementation gives:

```
<1>:
{X: 777,
 B: 
<2>:
    {B: -> <2>,
     A: -> <1>}}
```

while your implementation gives:

```
<#2> {X: 777,
      B: <#1> {B: <see #1>,
               A: <see #2>}}
```

Note that in your case definition for `<#2>` comes before definition for `<#1>`.

In my view with larger structures this can be harder to track and find a target if their definitions are not ordered. So if we want the targets to be ordered we need to sort them out somehow at writeout phase or just allocate their ids at the writeout - essentially what I did in my patch.

I also suggest to consider adjusting cosmetics a bit: put target definitions at newlines as this way it is easier to track them (imho) for a human and use `->` instead of `see` for references to raise signal/noise ratio.

Hope this helps a bit,
Kirill