@kylelemons 2017-08-23
Yep, you're right.  It would make the most sense if they were output in value order not reference order.  Looking at the code, I think I'd prefer to do a refactor of the WriteTo functions to keep the parameter list under control; I'm out for the rest of the week though, so I probably won't have a chance to get to it this weekend, but it should be pretty straightforward.

As for the horizontal spacing: I think I'm going to keep it the way I have it in my branch, just because that's the sort of style that Pretty has been using outside the diff context.  In particular, the change in indent layers is useful when diffs start to get more unruly.  For very deeply nested structures it will make them a bit wider, but not having the recursive context thing helps that significantly, but those are also the instances where having the reference flush-left will have it farthest from the value, so it's suboptimal in either case.

The last thing I noticed is that you enabled cycle tracking by default; unfortuantely, there are so many people using this package now that I'm very reluctant to change the default behavior (even performance wise).  It can actually be a significant hit depending on the structures you're using, and I anticipate things like protobuf that are very heavy on pointers will be the worst, but also not that likely to be self-referential.

Thanks for all the feedback!  I think this is coming out very nicely.