@navytux 2017-08-23
@kylelemons thanks for feedback. Good to hear we are converging on understanding what to do. About `WriteTo` refactor - you mainly mean that actual writeout state is needed only for labels/targets and not for other nodes?

spacing: ok;

protobuf: am I right you are meaning that `optional` fields are translated into pointers which are nil if option is not present? If so the following benchmark should simulate acyclic pointer-heavy data structure:

```diff
diff --git a/pretty/reflect_test.go b/pretty/reflect_test.go
index aa41238..15e89e6 100644
--- a/pretty/reflect_test.go
+++ b/pretty/reflect_test.go
@@ -275,6 +275,42 @@ func circular(nodes int) *ListNode {
        return recent
 }
 
+type Tree struct {
+       Nodev []*Node
+}
+
+type Node struct {
+       Leafv []*Leaf
+}
+
+type Leaf struct {
+       I, J int
+}
+
+// mktree creates new tree with nnode entries at level 1 and each level 1
+// having nleaf entries at level2.
+func mktree(nnode, nleaf int) *Tree {
+       tree := &Tree{}
+       for i := 0; i < nnode; i++ {
+               node := &Node{}
+               tree.Nodev = append(tree.Nodev, node)
+
+               for j := 0; j < nleaf; j++ {
+                       leaf := &Leaf{I: i, J: j}
+                       node.Leafv = append(node.Leafv, leaf)
+               }
+       }
+
+       return tree
+}
+
+// acyclic returns new config derived from cfg and .AssumeAcyclic set to true
+func acyclic(cfg *Config) *Config {
+       cfg2 := *cfg
+       cfg2.AssumeAcyclic = true
+       return &cfg2
+}
+
 func BenchmarkVal2node(b *testing.B) {
        benchmarks := []struct {
                desc string
@@ -283,12 +319,12 @@ func BenchmarkVal2node(b *testing.B) {
        }{
                {
                        desc: "acyclic/struct",
-                       cfg:  Acyclic,
+                       cfg:  acyclic(DefaultConfig),
                        raw:  struct{ Zaphod, Ford string }{"beeblebrox", "prefect"},
                },
                {
                        desc: "acyclic/map",
-                       cfg:  Acyclic,
+                       cfg:  acyclic(DefaultConfig),
                        raw: map[[2]int]string{
                                [2]int{-1, 2}: "school",
                                [2]int{0, 0}:  "origin",
@@ -324,6 +360,26 @@ func BenchmarkVal2node(b *testing.B) {
                        cfg:  DefaultConfig,
                        raw:  circular(3000),
                },
+               {
+                       desc: "tree(5,10)",
+                       cfg:  DefaultConfig,
+                       raw:  mktree(5, 10),
+               },
+               {
+                       desc: "acyclic/tree(5,10)",
+                       cfg:  acyclic(DefaultConfig),
+                       raw:  mktree(5, 10),
+               },
+               {
+                       desc: "tree(10,15)",
+                       cfg:  DefaultConfig,
+                       raw:  mktree(10, 15),
+               },
+               {
+                       desc: "acyclic/tree(10,15)",
+                       cfg:  acyclic(DefaultConfig),
+                       raw:  mktree(10, 15),
+               },
        }
 
        for _, bench := range benchmarks {
```

indeed for `tree(10,15)` case the difference is there:

```
name                            time/op
Val2node/acyclic/struct-4        506ns ± 0%
Val2node/acyclic/map-4          3.90µs ± 0%
Val2node/struct-4                608ns ± 1%
Val2node/map-4                  4.49µs ± 0%
Val2node/circlist/small-4       2.68µs ± 0%
Val2node/circlist/med-4          276µs ± 0%
Val2node/circlist/large-4       3.61ms ± 1%
Val2node/tree(5,10)-4           61.2µs ± 0%
Val2node/acyclic/tree(5,10)-4   51.7µs ± 0%
Val2node/tree(10,15)-4           174µs ± 0%
Val2node/acyclic/tree(10,15)-4   147µs ± 0%
```

i.e. `tree(10,15)` runs 174µs with cycle detection and 147µs without. However with patches like

```diff
diff --git a/pretty/reflect.go b/pretty/reflect.go
index 34e4d60..0f69a55 100644
--- a/pretty/reflect.go
+++ b/pretty/reflect.go
@@ -19,6 +19,7 @@ import (
        "fmt"
        "reflect"
        "sort"
+       "strconv"
 )
 
 func isZeroVal(val reflect.Value) bool {
@@ -158,9 +159,11 @@ func (ref *reflector) val2node(val reflect.Value) node {
                }
                return rawVal("false")
        case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-               return rawVal(fmt.Sprintf("%d", val.Int()))
+               //return rawVal(fmt.Sprintf("%d", val.Int()))
+               return rawVal(strconv.FormatInt(val.Int(), 10))
        case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64:
-               return rawVal(fmt.Sprintf("%d", val.Uint()))
+               //return rawVal(fmt.Sprintf("%d", val.Uint()))
+               return rawVal(strconv.FormatUint(val.Uint(), 10))
        case reflect.Uintptr:
                return rawVal(fmt.Sprintf("0x%X", val.Uint()))
        case reflect.Float32, reflect.Float64:
```

we can easily with that back:

```
name                            old time/op    new time/op    delta
Val2node/acyclic/struct-4          506ns ± 0%     501ns ± 1%   -0.91%  (p=0.000 n=9+10)
Val2node/acyclic/map-4            3.90µs ± 0%    3.24µs ± 0%  -16.93%  (p=0.000 n=9+10)
Val2node/struct-4                  608ns ± 1%     600ns ± 0%   -1.30%  (p=0.000 n=10+10)
Val2node/map-4                    4.49µs ± 0%    3.80µs ± 0%  -15.53%  (p=0.000 n=9+8)
Val2node/circlist/small-4         2.68µs ± 0%    2.29µs ± 0%  -14.49%  (p=0.000 n=9+10)
Val2node/circlist/med-4            276µs ± 0%     241µs ± 1%  -12.56%  (p=0.000 n=10+10)
Val2node/circlist/large-4         3.61ms ± 1%    3.03ms ± 1%  -15.88%  (p=0.000 n=10+10)
Val2node/tree(5,10)-4             61.2µs ± 0%    50.2µs ± 0%  -18.07%  (p=0.000 n=10+10)
Val2node/acyclic/tree(5,10)-4     51.7µs ± 0%    41.3µs ± 1%  -20.15%  (p=0.000 n=9+10)
Val2node/tree(10,15)-4             174µs ± 0%     142µs ± 0%  -18.36%  (p=0.000 n=10+10)
Val2node/acyclic/tree(10,15)-4     147µs ± 0%     117µs ± 0%  -20.51%  (p=0.000 n=10+8)

name                            old alloc/op   new alloc/op   delta
Val2node/acyclic/struct-4           176B ± 0%      176B ± 0%     ~     (all equal)
Val2node/acyclic/map-4            1.50kB ± 0%    1.44kB ± 0%   -3.48%  (p=0.000 n=10+10)
Val2node/struct-4                   224B ± 0%      224B ± 0%     ~     (all equal)
Val2node/map-4                    1.69kB ± 0%    1.64kB ± 0%   -3.08%  (p=0.000 n=10+10)
Val2node/circlist/small-4           832B ± 0%      784B ± 0%   -5.77%  (p=0.000 n=10+10)
Val2node/circlist/med-4           81.4kB ± 0%    78.2kB ± 0%   -3.95%  (p=0.000 n=10+10)
Val2node/circlist/large-4          773kB ± 0%     748kB ± 0%   -3.23%  (p=0.000 n=10+10)
Val2node/tree(5,10)-4             15.8kB ± 0%    14.3kB ± 0%   -9.38%  (p=0.000 n=10+10)
Val2node/acyclic/tree(5,10)-4     15.6kB ± 0%    14.1kB ± 0%   -9.50%  (p=0.000 n=10+10)
Val2node/tree(10,15)-4            43.3kB ± 0%    38.7kB ± 0%  -10.64%  (p=0.000 n=10+10)
Val2node/acyclic/tree(10,15)-4    43.1kB ± 0%    38.5kB ± 0%  -10.69%  (p=0.000 n=10+10)

name                            old allocs/op  new allocs/op  delta
Val2node/acyclic/struct-4           7.00 ± 0%      7.00 ± 0%     ~     (all equal)
Val2node/acyclic/map-4              53.0 ± 0%      44.0 ± 0%  -16.98%  (p=0.000 n=10+10)
Val2node/struct-4                   8.00 ± 0%      8.00 ± 0%     ~     (all equal)
Val2node/map-4                      55.0 ± 0%      46.0 ± 0%  -16.36%  (p=0.000 n=10+10)
Val2node/circlist/small-4           34.0 ± 0%      28.0 ± 0%  -17.65%  (p=0.000 n=10+10)
Val2node/circlist/med-4            3.02k ± 0%     2.62k ± 0%  -13.23%  (p=0.000 n=10+10)
Val2node/circlist/large-4          30.1k ± 0%     27.0k ± 0%  -10.31%  (p=0.000 n=10+10)
Val2node/tree(5,10)-4                752 ± 0%       567 ± 0%  -24.60%  (p=0.000 n=10+10)
Val2node/acyclic/tree(5,10)-4        750 ± 0%       565 ± 0%  -24.67%  (p=0.000 n=10+10)
Val2node/tree(10,15)-4             2.20k ± 0%     1.62k ± 0%  -26.16%  (p=0.000 n=10+10)
Val2node/acyclic/tree(10,15)-4     2.20k ± 0%     1.62k ± 0%  -26.18%  (p=0.000 n=10+10)
```

and not only for large tree cases - i.e. in several cases cyclic + speedup becomes actually faster then old acyclic without speedup.

This way I'm suggesting to apply the speedup patch and change defaults to be cycles are detected, not crashed with.

Kirill

P.S. Once https://github.com/golang/go/issues/19495 is hopefully fixed tracking pointer should become cheaper.